Move test map function and add parent map entity
Group all map entities under a parent entity so that the map can
more easily be unloaded and loaded in the future.
